Add branching capabilities to deposition
Define a ['development']['branch'] object in the yaml config

By default, the rundir of the simulations will the one of the deposition
number with the branch attribute appended to it.
NOTABLE BUG : Dont use numbers for branches, or expect clashes

To branch a deposition from the default one, simply run it so that the
starting simulation does not currently exist on your branch, then by
default it will source the default '' one.
NB : THe behaviour could be rationalized a lot more.